requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.0
urllib3==2.6.3
